+++ .gitignore
@@ -16,4 +16,4 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-.DS_Store
+*.DS_Store
