--- globalStyles.js
@@ -1,6 +1,0 @@
-export const NAVY = '#072956';
-export const BLUE = '#527edd';
-export const CREAM = '#f5f1ed';
-export const FONT_NORMAL = 'Montserrat-Regular';
-export const FONT_BOLD = 'OpenSans-Bold';
-export const BORDER_RADIUS = 5;
