+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "test2",
+  "name": "bmu-client",
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
@@ -17,13 +17,14 @@
   },
   "jest": {
     "preset": "jest-expo"
+    
   },
   "dependencies": {
     "expo": "^23.0.4",
     "moment": "^2.20.1",
     "react": "16.0.0",
     "react-moment": "^0.7.0",
-    "react-native": "0.50.3",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-23.0.0.tar.gz",
     "react-native-card-list": "^0.1.0",
     "react-native-datepicker": "^1.6.0",
     "react-native-elements": "^0.19.0",
