+++ src/constants/styles.js
@@ -3,3 +3,9 @@ import { Alert, DatePickerIOS, StyleSheet, Text, View, AppRegistry, Image, TextI
 import { StackNavigator, } from 'react-navigation';
 import { List, ListItem, Button } from 'react-native-elements'; 
 
+export const NAVY = '#072956';
+export const BLUE = '#527edd';
+export const CREAM = '#f5f1ed';
+export const FONT_NORMAL = 'Montserrat-Regular';
+export const FONT_BOLD = 'OpenSans-Bold';
+export const BORDER_RADIUS = 5;
